when removing user, user is also removed from company

+++ server/schemas/resolvers.js
@@ -296,12 +296,42 @@ const resolvers = {
             }
         },
 
-        removeUser: async (parent, { userId }) => {
+        // removeUser: async (parent, { userId }, context) => {
+        //     if (
+        //         context.user.role == "Owner" ||
+        //         context.user.role == "Admin"
+        //     ) {
+        //         const user = await User.findByIdAndDelete(userId);
+        //         return user;
+        //     } else {
+        //         throw new AuthenticationError(
+        //             "You need to be logged in and be an owner/admin to delete a user"
+        //         );
+        //     }
+        // },
+        removeUser: async (parent, { userId }, context) => {
             if (
                 context.user.role == "Owner" ||
                 context.user.role == "Admin"
             ) {
-                const user = await User.findByIdAndDelete(userId);
+                const user = await User.findById(userId);
+
+                 if (!user) {
+                    throw new Error("User not found");
+                 }
+                
+                await User.findByIdAndDelete(userId);
+
+                if (user.company) {
+                  const company = await Company.findById(user.company);
+
+                  if (company) {
+                    company.users.pull(userId);
+                    await company.save();
+                  }
+                }
+
+                
                 return user;
             } else {
                 throw new AuthenticationError(
